requirements.txt
discord.py
google.generativeai
generativeai